requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --extra=llm --extra=local --output-file=requirements.txt pyproject.toml
#
accelerate==1.3.0
    # via aho (pyproject.toml)
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via aho (pyproject.toml)
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anthropic==0.45.2
    # via aho (pyproject.toml)
anyio==4.8.0
    # via
    #   anthropic
    #   groq
    #   httpx
    #   openai
async-timeout==5.0.1
    # via aiohttp
attrs==25.1.0
    # via aiohttp
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
distro==1.9.0
    # via
    #   anthropic
    #   groq
    #   openai
exceptiongroup==1.2.2
    # via anyio
filelock==3.17.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.12.0
    # via
    #   huggingface-hub
    #   torch
groq==0.15.0
    # via aho (pyproject.toml)
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   aho (pyproject.toml)
    #   anthropic
    #   groq
    #   openai
huggingface-hub==0.28.0
    # via
    #   accelerate
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.5
    # via torch
jiter==0.8.2
    # via
    #   anthropic
    #   openai
joblib==1.4.2
    # via scikit-learn
loguru==0.7.3
    # via aho (pyproject.toml)
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
networkx==3.4.2
    # via
    #   aho (pyproject.toml)
    #   torch
numpy==2.2.2
    # via
    #   accelerate
    #   aho (pyproject.toml)
    #   scikit-learn
    #   scipy
    #   transformers
openai==1.60.2
    # via aho (pyproject.toml)
packaging==24.2
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
pillow==11.1.0
    # via sentence-transformers
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
psutil==6.1.1
    # via accelerate
pydantic==2.10.6
    # via
    #   aho (pyproject.toml)
    #   anthropic
    #   groq
    #   openai
pydantic-core==2.27.2
    # via pydantic
python-dotenv==1.0.1
    # via aho (pyproject.toml)
pyyaml==6.0.2
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   transformers
safetensors==0.5.2
    # via
    #   accelerate
    #   transformers
scikit-learn==1.6.1
    # via
    #   aho (pyproject.toml)
    #   sentence-transformers
scipy==1.15.1
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==3.4.0
    # via aho (pyproject.toml)
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   groq
    #   openai
sympy==1.13.1
    # via torch
tenacity==9.0.0
    # via aho (pyproject.toml)
threadpoolctl==3.5.0
    # via scikit-learn
tokenizers==0.21.0
    # via transformers
torch==2.5.1
    # via
    #   accelerate
    #   aho (pyproject.toml)
    #   sentence-transformers
tqdm==4.67.1
    # via
    #   aho (pyproject.toml)
    #   huggingface-hub
    #   openai
    #   sentence-transformers
    #   transformers
transformers==4.48.1
    # via
    #   aho (pyproject.toml)
    #   sentence-transformers
typing-extensions==4.12.2
    # via
    #   aho (pyproject.toml)
    #   anthropic
    #   anyio
    #   groq
    #   huggingface-hub
    #   multidict
    #   openai
    #   pydantic
    #   pydantic-core
    #   torch
urllib3==2.3.0
    # via requests
yarl==1.18.3
    # via aiohttp
